add `NeverDone` function for package `gctx` that creates and returns a never done ctx (#2784)
