+++ contrib/drivers/oracle/oracle.go
@@ -7,9 +7,8 @@
 // Package oracle implements gdb.Driver, which supports operations for database Oracle.
 //
 // Note:
-// 1. It needs manually import: _ "github.com/sijms/go-ora/v2"
-// 2. It does not support Save/Replace features.
-// 3. It does not support LastInsertId.
+// 1. It does not support Save/Replace features.
+// 2. It does not support LastInsertId.
 package oracle
 
 import (
@@ -19,7 +18,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/gogf/gf/v2/util/gutil"
 	gora "github.com/sijms/go-ora/v2"
 
 	"github.com/gogf/gf/v2/database/gdb"
@@ -28,6 +26,7 @@ import (
 	"github.com/gogf/gf/v2/text/gregex"
 	"github.com/gogf/gf/v2/text/gstr"
 	"github.com/gogf/gf/v2/util/gconv"
+	"github.com/gogf/gf/v2/util/gutil"
 )
 
 // Driver is the driver for oracle database.
